Support code generation for parameterized imports
This adds support to `pulumi import` for generating code for resources
from parameterized packages.

This requires two things. Firstly we need the whole state when
generating code so we can look up providers (there was a very old `TODO`
comment that we should have been doing this anyway so we used the
correct package versions). Looking up the provider means we can pull all
the package information off it to look up the schema correctly.

We also emit that package information to the generated PCL as package
blocks, so that the language code generator also has full access to the
package information so it can correctly load schemas inside its
`GenerateProgram` call.

Part of #17507