Run Python conformance tests in parallel (#18216)
Presently, conformance tests for Python run in serial, due to various
historic issues with shared state and the like. It seems that these have
now been fixed, so with some rejigging and appropriate `t.Run` nesting,
we should now be able to run the tests in parallel (as we do for e.g.
Node). This change does just that.

Part of #13945