Fix callbackserver shutdown (#18190)
When using transforms, we spin up a GRPC server within the Python
runtime to run the transforms. During shutdown of the Python runtime,
this server would print errors when running in Python 3.12 or higher.

By using the
[wait_for_termination](https://grpc.github.io/grpc/python/grpc_asyncio.html#grpc.aio.Server.wait_for_termination)
instead of
[stop](https://grpc.github.io/grpc/python/grpc_asyncio.html#grpc.aio.Server.stop)
we avoid these errors.

Fixes https://github.com/pulumi/pulumi/issues/18176