Add the `l2-component-call-simple` conformance test (#18207)
This change exercises PCL's new `call` intrinsic to introduce a
conformance test for testing code generation and execution for "simple"
applications of the `Call` method on the `ResourceMonitor` gRPC
interface (and behind it, the `ResourceProvider` interface). As part of
this, we extend programgen for NodeJS and Python, where it's a
relatively simple adaptation of the existing `invoke` machinery. It
seems that Go will require a bit more work due to some errant name
generation, which is tracked by #18202.

Later on we can add tests for more complex uses of `Call`, such as on
rehydrated resource references, or perhaps using "multi-argument
signatures", but this should suffice for now.

Part of fixing pulumi/pulumi-java#262