Support parameterized packages in `pulumi import` (#18038)
The `pulumi import` command allows users to import existing provider
resources into a Pulumi stack, accepting the type of resource to import,
a name for the resource that will be created in the stack, and some "ID"
that identifies the existing resource in the provider. Historically, the
type has been sufficient to infer which provider offers the resource,
since package names were in a one-to-one correspondence with plugin
names. Now that we support parameterized providers, however, this is no
longer the case -- we might wish to import a `random:index/id:Id`, but
where the `random` package is a dynamically-bridged instance of the
Terraform `random` provider produced by the `terraform-provider` plugin.

This change makes `import` aware of parameterizations by letting it deal
with fully-specified package descriptors, as opposed to just plugin
specifications. In doing so, `import` can now support importing
parameterized resources. That said, it still does not support code
generation for such imports (that is, `--generate-code` must be set to
`false`) -- this will come in a later change.

Part of #17507