Add continue on error to pulumi preview.
This also will cause the CLI to exit with code 0 when this flag is set,
even if there is an error deleting a resource due to failed
performDelete of a resource.

This is to allow for CI/CD pipelines to work correctly for protected
resources when there is no actual change in a preview.