Retrieve Git metadata from the environment (#12871)
When running a Pulumi operation, we attempt to gather various bits of
metadata, such as information about the commit, committer, author, and
so on from the available version control system (VCS). In some
situations, Pulumi is being run from a context where it doesn't have
access to e.g. a `.git` folder, and so this information isn't added.
This change allows this information to be added explicitly by setting a
set of environment variables, enabling functionality that hangs off this
metadata (e.g. that of the Pulumi GitHub action) to work correctly in
these cases. The variables in question are:

* `PULUMI_VCS_REPO_OWNER`: The owner of the repository
* `PULUMI_VCS_REPO_NAME`: The name of the repository
* `PULUMI_VCS_REPO_KIND`: The kind of repository (e.g. `github.com`)
* `PULUMI_VCS_REPO_ROOT`: The repository path where Pulumi is being run
* `PULUMI_GIT_HEAD_NAME`: The name of the Git head being used
* `PULUMI_GIT_HEAD`: The SHA hash of the Git head being used
* `PULUMI_GIT_COMMIT_MESSAGE`: The commit message
* `PULUMI_GIT_COMMITTER`: The name of the committer
* `PULUMI_GIT_COMMITTER_EMAIL`: The email of the committer
* `PULUMI_GIT_AUTHOR`: The name of the commit author
* `PULUMI_GIT_AUTHOR_EMAIL`: The email address of the commit author

Co-authored-by: Will Jones <will@sacharissa.co.uk>