Allow resource annotations to be publicly queried (#18276)
In #18206 we introduced the `call` intrinsic to PCL. As part of making
this work, we improved the annotating of PCL expressions with their
schema resource types, allowing us to e.g. type-check `call` expressions
by traversing the schema attached to a receiver expression. The type we
introduced to support these annotations, `resourceAnnotation`, was
initially made private, since it was only used within the package.
However, in now trying to add `call` to e.g. Java, it's become apparent
that it would be valuable to access the annotation from other packages.
This change thus makes it public.

The other option here is to always wrap the receiver ("self") in a
`__convert` that binds its schema type, but since the annotation is
already attached this seems simpler and a bit cleaner. In the long term,
we'd ideally tweak how we do code generation so that schema types are
readily available throughout the various generation method calls, but
right now this isn't currently the case.