allow setting the type of the config value in pulumi config set
Config values in `pulumi config set` are marshalled according to
certain rules (everything on the top level is a string), and for
paths, true/false becomes a boolean, integers not starting with a 0
become integers and everything else becomes strings.

This leaves floats out, and someone might want to store a number or a
bool as a string instead.  Add a `--type` flag to pulumi config set to
force the type.  This type also applies to top level config elements.