Fix TestListPackages (#18292)
The latest version of pulumi-random requires Python >= 3.9, but the
generated pyproject.toml files claim the project should run with >= 3.8,
leading to unresolvable deps.

Because the latest version of `pulumi-random` now also pulls in the
version of `pulumi` that has an explicit dependency on `pip`, we need to
adjust the code for that.

As a follow up, we should update this test to not use `pulumi-random`,
but a mock package instead, to avoid similar issues on the next bump of
the minimum support Python version.