Allow resource annotations to be publicly queried, for real
This is a follow-up to #18276, in which we made the `ResourceAnnotation`
type public, but forgot to also make its `node` field public at the same
time. Doing this should actually make the type usefully accessible to
other packages, so that we can proceed with e.g. implementing `Call` in
Java as intended.