Add some skeletal documentation on the built-in provider
As part of the recently added conformance tests for testing components,
we now have better coverage (and a better understanding of) resource
references and their hydration, which is handled by the built-in
provider. This small change makes a start at writing some of this down,
adding a section for the built-in provider and adjusting the
documentation we have on resource references to reference it.