Permit output arguments in plain invokes for backwards compatibility
Plain invokes happen to serialize arguments, which allows passing in output types. With https://github.com/pulumi/pulumi/pull/18204 we started checking if the dependency resources of the arguments (and via `dependsOn`) are known, and returned unknown if not. However this broke backwards compat for plain invokes in the TypeScript and Python SDKs.

This PR ensures that we only check the dependencies if we are using output form invokes.