Avoid unnecessary Node.js installations (#18293)
When running `pulumi install --use-language-version-tools`, and `fnm` is
installed on the system, we will automatically install the Node.js
version specified in `.nvmrc` or `.node-version`.

When the file does not specify the full version, for example `22`, we
would always install the latest release of Node.js 22, even if another
point release of Node.js 22 is already present on the system. This is
unnecessary, and we should use the available version.

This mostly affects Pulumi Deployments, where this installation flag is
used by default. Whenever we release a new version of Pulumi, we also
release new versions of the docker containers. These containers include
the latest version of Node.js (22, and other versions) at the time of
release, so the pre-installed version is usually very recent. However if
a new Node.js version is released after our release, deployments that
specify a Node.js version will always install that, until we update the
containers with the next Pulumi release.

By using `fnm use ${VERSION} --install-if-missing` instead, we prefer
pre-installed versions, and only install new Node.js versions, if the
request version can not be satisfied by any of the pre-installed
versions.

https://github.com/pulumi/pulumi/pull/18041 was a previous attempt to
fix this, but it did not properly handle the fnm shell integration and
had to be reverted.

Ref https://github.com/pulumi/customer-support/issues/1863