Permit output arguments in plain invokes for backwards compatibility (#18304)
Plain invokes happen to serialize arguments, which allows passing in
output types. With https://github.com/pulumi/pulumi/pull/18204 we
started checking if the resource dependencies of the arguments (and from
`dependsOn`) are known, and returned unknown if not. However this broke
backwards compatibility for plain invokes in the TypeScript and Python
SDKs.

This PR ensures that we only check the dependencies if we are using
output form invokes. In Python and TypesScript, arguments of output type
decay to promises/tasks, without dependency information, when passed to
plain invokes.