allow installing plugins from subdirectories
We want to allow plugins to exist at subdirectories of a repository,
so there can be a monorepo with multiple plugins.  At the same time,
those plugins might depend on something in the repository, so we can't
just naively copy a sub-part of the repo into `.pulumi/plugins`,
because the plugin might depend on something in a parent directory.

To solve this, we always install the whole repo into
`.pulumi/plugins`, but the allow referring to the path within that
repository.  This works well with the current design, as all we really
need to do is point the local name to the right directory and it works.