When loading schema, query plugin info for version if it is paramaterized.
Not doing this is causing package references to not be loaded during binding of
PCL and discarded, causing the go language info structure to not be available from the package.

This in turn means that the import codegen logic is incorrect and
produces the wrong path.  This solves the version mismatch error problem
so that does not error out and cause the bad effects.

Fixes #18310