automatically get latest version for plugins from a git source (#18215)
Plugins from a git source are being introduced in
https://github.com/pulumi/pulumi/pull/17827. That implementation
requires the user to specify the version manually. We can and want to do
better however. This commit makes it so that we detect the version
automatically from the repo.

We do this by first checking if there are any tags that parse as semver,
and if so return the latest tag. If there are no tags that are valid
semver versions, we use the commit hash for the latest commit on the
default branch.

Fixes https://github.com/pulumi/pulumi/issues/18199
Depends on https://github.com/pulumi/pulumi/pull/17827