add test infrastructure for `pulumi config set` (#18305)
Add test infrastructure to test the `pulumi config set` command, and add
a simple test for it. This test isn't meant to comprehensively test
everything in the command, but rather as a start to be able to add more
tests more easily as we go and add more features.

I wanted to add tests for https://github.com/pulumi/pulumi/pull/18287,
but without the test infrastructure in place it's rather painful. This
is a step towards that, so #18287 can be tested properly.