Convert PEP440 local version to semver prerelease in setup.py (#18289)
PEP440 only allows a small set of prerelease tags, so when converting an
arbitrary prerelease, PypiVersion in `/pkg/codegen/python/utilities.go`
converts it to a "local" version with a `+` sign. This PR does the
reverse conversion and sets the local version as the prerelease tag.

Resolve https://github.com/pulumi/pulumi/issues/18288