Preview conformance tests (#18295)
### Description 

This PR adds a preview step to every conformance test with the ability
to assert certain aspects of the resource changes and the generated plan
via an `AssertPreview` function. When the tests do not specify an
`AssertPreview` function, we give a default implementation that simply
checks whether the preview is successful.

The PR makes slight changes to a couple conformance provider
implementations such that they don't hard fail when encountering unknown
values (during preview)

Fixes #18296

---------

Co-authored-by: Will Jones <will@sacharissa.co.uk>