Fix nesting directories for SDK conversion
Currently to generate and link sdks, we change into the output directory
so the underlying functionality which depends on the workspace functions
as expected, but we forget to pop out of this directory every iteration
of the loop so we end up nesting unexpectedly.

This change gets rid of this global state and instead uses paths
directly.

Fixes #18312