Allow pip 25 (#18343)
Pip 25 was released and this version works fine, allow it in our
dependencies.

This should avoid dependency installation getting confused when trying
to upgrade pip in a venv, which has lead to some test flakiness. Note
that this seems to happen due to parallel tests using the global
evnironment, and does not seem to be an issue in normal usage.