Fix overlap when "this" keyword used in ts generation. (#18177)
As other keyword overlaps come up we can add them to this switch
statement.

This change also includes fixes to the go generation to make these tests
pass. Previously we were chekcing the wrong identifiers against the
roots in the pcl program. Now we use the original names as intended and
keywords work correctly now.

Fixes #18348