Make RunPlugin default to using a virtualenv (#18371)
For Python plugins `PluginSpec.InstallDependencies` always installs
plugin dependencies into a virtual environment named `venv` in the
plugin's directory.

Python's `RunPlugin` reads the options from `PulumiPlugin.yaml` to
determine the toolchain to use. If no toolchain options are set, this
proceeds to use the dependencies from the global Python installation,
which is not what we want to happen. Instead, we should default to
`toolchain: pip` (the default already) with `virtualenv: venv`, to match
the installation step.

Fixes https://github.com/pulumi/pulumi/issues/18347