align pulumi.json and schema.go (#18334)
As part of https://github.com/pulumi/pulumi/issues/18203, I wanted to
first go through `pulumi.json` and `schema.go` and see if they match.
They do mostly, but not completely. This pull request brings them closer
in sync. I opted for adding things here rather than removing them, as
that's not going to cause compatibility problems.

Note that objectTypeSpec doesn't quite match the schema still, however
it matches it "in spirit". It has additional fields in the Go code,
which are only defined on the objects that embed it. I've left that as
it is for now.

(This is my first time touching the schema, so please let me know if I'm
missing anything here, or if this doesn't make sense)