add some documentation for appending .git to the URL to disambiguate (#18379)
We support this disambiguation mechanism, but it's not widely known, and
I don't think it's documented anywhere. Add it to the documentation for
`pulumi new` to publicise this feature a bit more widely.

Fixes https://github.com/pulumi/pulumi/issues/17582 (I don't think we
can do much better here, but happy to leave the issue open if others
disagree)