Use the corepack executable from dev dependencies
The corepack version that ship with current nodejs releases include expired signatures, causing the installation of the latest pnpm and npm versions to fail. Instead of relying on an implicit corepack version, include a version with recent signatures (>=0.31.0) in our dev dependencies, and use that in our tests.

Fixes https://github.com/pulumi/pulumi/issues/18338