Use the corepack executable from dev dependencies (#18393)
The corepack version that ships with current nodejs releases uses old
keys to verify signatures, causing the installation of the latest pnpm
and npm versions, which use a newer key, to fail. Instead of relying on
an implicit corepack version, include a version with recent signatures
(>=0.31.0) in our dev dependencies, and use that in our tests.

Fixes https://github.com/pulumi/pulumi/issues/18338