Fix some plugin context leaks (#18373)
Every call of `plugin.NewContext` needs to have a matching `Close`.
`ProviderFromSource` was creating a context and never closing it leading
to process leaks from commands like `gen-sdk`. I'm not sure this fixes
all the context leaks, I didn't search the whole codebase but it fixes
everything related to `ProviderFromSource`.

Fix is simple that we need to push the context creation up to cover the
whole lifecycle of the provider being used. This is generally the top
level function, where we just defer the context close to ensure it's
shut down on command exit.

---------

Co-authored-by: Thomas Gummerer <t.gummerer@gmail.com>