Fix parsing of config keys with paths and colons in path segments (#18378)
Currently, the following command fails:

```
$ pulumi config set pulumi:autonaming.providers.aws.resources.aws:s3/bucket:Bucket.mode verbatim --path=true
error: invalid configuration key: could not parse pulumi:autonaming.providers.aws.resources.aws:s3/bucket:Bucket.mode as a configuration key (configuration keys should be of the form `<namespace>:<name>`)
```

because we expect 0 to 2 `:` symbols in the configuration key string.

This expectation is too strict for the `path: true` case, because colons
may appear in non-top path segments, as in the above example. The
command above should produce

```
config:
  pulumi:autonaming:
    providers:
      aws:
        resources:
          aws:s3/bucket:Bucket:
            mode: verbatim
```

This PR modifies `ParseConfigKey` behavior to allow that.
`ParseConfigKey` now accepts whether the given key is supposed to be a
path. If it's not a path, the behavior is the same as before. If it is a
path, we attempt to parse the key as path, and then only apply the
namespace calculation to the top-level segment (e.g. auto-prefix it with
the project name if needed). We then use that calculated namespace to
construct the full Key.

One pesky detail is that the old command allowed some keys that are not
valid paths, e.g. `pulumi config set aws:boo. value1 --path=true`
succeeded before but would fail if we require it to be a valid path. To
keep backwards compat, I chose to fallback to non-path parsing if path
parsing fails. It's ugly but avoids breaking existing users...

After the change, `pulumi config set
pulumi:autonaming.providers.aws.resources.aws:s3/bucket:Bucket.mode
verbatim --path=true` works as expected.

I'd love to get a careful review on what else I may be breaking or other
test cases that I'm missing.

Resolve https://github.com/pulumi/pulumi/issues/18342