fix flaky TestRunCanceled (#18390)
This test currently flakes sometimes because the stdiout is "signal:
interrupt" instead of "error: preview canceled". This can happen
depending on when exactly the preview is canceled.

It would be nicer if the error always said "preview canceled", but the
important thing is that we interrupt the preview successfully, and it's
probably not worth our time right now to figure out exactly where this
is flaky and why we're not getting the nice output in some rare cases.

It is however important that the test is not flaky, so let's just fix
the test to also accept "signal: interrupt" as potential output when the
preview is canceled.

Fixes https://github.com/pulumi/pulumi/issues/18389