implement GetPackageInfo
This is the first of what's going to be a series of pull requests to
allow getting partial info about the schema from a provider.  The idea
here is that instead of a JSON schema that we need to parse on the
receiver side, and have to stitch together on the provider side, we
can use proto types to send this data.

At the same time it also starts allowing getting partial information
from the provider, potentially cutting down on the traffic we send
over the gRPC interface.

This PR implements getting the basic PackageInfo.  Following PRs will
implement getting resources and types in a similar fashion, which is
all the data we need for components-as-components.