python: import packages from <publisher>_
Curently when using `package add`, packages are always namespaced in
the "pulumi" namespace.  For python this means they are always named
`pulumi_<name>`.  However it would be better to namespace them by
publisher if that was given, so we can disambiguate them better.  This
PR makes that so for Python.  Subsequent PRs are going to implement
the same thing for other languages.