Stream standard output and error when installing dependencies (#18404)
The `InstallDependencies` method of the language host gRPC interface
supports streaming back data for the standard output and error of the
processes being executed by the language host (e.g. `npm install` for
NodeJS). Presently however, we do not expose these streams in the
higher-level `LanguageHost` Go interface. This commit changes that,
having `LanguageHost`'s `InstallDependencies` method return a pair of
`io.Reader`s for standard output and error respectively, as well as a
channel that will be closed when the operation completes. This
transforms `InstallDependencies` to an asynchronous call and so call
sites are modified appropriately to drain the readers and block on the
`done` channel's closure. With this change, we can modify the language
conformance test server to capture standard output and error when
dependency installation goes wrong, making it easier to debug
conformance tests in the process.

Fixes #13941