allow integration tests to automatically rebuild binaries
One of the most common pitfalls for me is forgetting to rebuild the
binaries before running the integration tests.  There's currently no
automated way to do it, and since the integration tests just pick up
the binaries from $PATH they just use whatever is currently installed
on the system.

This PR introduces a new environment variable that can be set, that
makes the integration tests automatically build the binary into a
`bin` folder in the local repository.  This folder is then added to
the path, so `pulumi` and the language hosts from that path are being
used instead of the system version.