Filter __internal keys when calling provider diff (#18440)
Noticed this while looking into
https://github.com/pulumi/pulumi/issues/18449.

Whenever we call provider methods to do with config (e.g. DiffConfig) we
should filter out the "__internal" key that the engine adds to state for
tracking engine internal data. We were doing this correctly everywhere
except in one place in the step generator when calculating diffs due to
provider changes.

This fixes that place up and also adds a test to check we don't regress
this.