Use old inputs for Same steps (#18411)
Prompted by https://github.com/pulumi/pulumi-docker/issues/1339

As part of investigating this issue it transpired that the engine
currently uses the new inputs for provider configuration, even if the
provider says there's no difference between the old inputs and the new
inputs.

This leads to inconsistencies when a provider reports no diff, but the
values have actually changed. Worse this leads to inconsistencies of if
the new or old inputs are used based on the order of operations.

In a normal update the new inputs from the program (or config) would be
used to configure the provider, but in a refresh update (e.g. `up
--refresh`) the engine would first configure the provider based on the
old inputs saved in state. As diff then reported "no diff" the engine
would never re-initialize and configure with the new inputs. So
depending on if you added `--refresh` or not you either got new or old
configuration for the provider.

This makes it consistent that we _always_ use the old inputs in state if
the provider reports that there is no configuration diff change.

N.B. We still _save_ the new inputs into state, and so on the next run
_they_ will be the old inputs. This matches what we do for same steps
for other resources. It feels like there might be an inconsistency
lurking here as well, but that needs more thinking and discussion before
actioning.