[Experimental] Handle inputs and outputs (#18408)
Names need to be camel cased in the schema. Going from snake_case to
camelCase is fine, however for the opposite direction we need to keep
track of the original names. For example `URL` might have been `u_r_l`
or `URL` originally.

With the mapping in place, we can send inputs into the component, and
retrive the outputs.

Ref https://github.com/pulumi/pulumi/issues/18366