Always print summaries in preview (#18465)
This changes the CLI to always print the resource summary at the end of
previews. Like `up` in the case of errors this might not match what the
program _wants_ to do. But errors in preview are unusual, the main one
being due to trying to delete protected resources.

Users have requested that they still want to see the summary if the only
error is protect deletes (see
https://github.com/pulumi/pulumi/issues/18103).

Fixes https://github.com/pulumi/pulumi/issues/18103.

---------

Co-authored-by: Thomas Gummerer <t.gummerer@gmail.com>