Add PulumiHomeDir to ProgramTestOptions (#18477)
This allows tests to set the PULUMI_HOME directory, rather than always
using a new temp directory. Mostly intended for benchmarking where we
want a prewam run of `ProgramTest` followed by the real run but both
need to use the same PULUMI_HOME so that plugins only get downloaded on
the prewarm run.