Rejig import files to record base plugin names and versions in the paramterisation block (#18450)
An alternative take on https://github.com/pulumi/pulumi/pull/18259

Noted while looking at that PR that we were duplicating information into
the import file (the parameterized package name, was already part of the
type token, and was being declared in the parameterization block).

Parameterization is awkward because in some places we have the plugin
name at the root level (e.g. for package descriptors) and in other
places the plugin name is in the parameterisation block (e.g. in
schema).

This results in quite a lot of book keeping of when we need to flip from
one to the other. Imports are another case of this.
The import file already defines "type" which specifies the package name
as part of that. So it needs the base plugin name as extra information
in the parameterisation block.
But import also deals with default provider information which is made of
package descriptors which have the plugin name at the root level, not
the package name.
Thus these required inverting transforms.

Closes #18259