Swap BulkDecrypt to return list (#18486)
In preparation for introducing a `BulkEncrypt([]string) []string` where
inputs and outputs are correlated based on order, it will also be good
to use the same pattern of access for the existing decrypt methods. This
does not yet change the pattern of how these values are returned by the
service for bulk decryption - only how the library functions operate.

Most critical changes:
- mapDecrypter.BulkDecrypt (pkg/resource/stack/secrets.go) - added
missing test coverage here.
- serviceCrypter.BulkDecrypt (pkg/secrets/service/manager.go) - already
covered by existing tests.

This changes the interface within sdk/go/common/resource/config but this
does not appear to be implemented outside of pulumi/pkg based on GitHub
search.