package add: avoid panic when Pulumi.yaml is invalid
In `pulumi package add`, we first read the project, and then use that
information.  There is some error checking there, however we only
error out if the project is not found, and not for other errors
reading the project.

Continuing when there is an error reading the project just leads to
potential panics down the line.  E.g. if the project isn't read
successfully, we panic in the next lines trying to get the name of the
project.

Just error out consistently if there is an error reading the project
instead, so the user gets a better hint of what is wrong and why the
command is not working rather than a panic.