[Experimental/Components]: Infer type and property descriptions (#18540)
The type and property descriptions are inferred from the docstrings of
the
Python classes and their attributes. For classes, the docstrings are
available on `type.__doc__`. Unfortunately, the docstrings of the class
attributes are not present at runtime. We have to parse the source code
to
retrieve them, which means we are doing two passes over all the Python
files
in the module: one pass of runtime inspection to determine the types,
and a
second pass of source code parsing to get the docstrings.

Fixes https://github.com/pulumi/pulumi/issues/18438
Ref https://github.com/pulumi/pulumi/issues/15939