Conformance test that takes dynamic list input and returns dynamic values (#18424)
In an effort to make conversion more robust (#18346), some pulumi-std
functions will be enabled for conversion from terraform code.

These functions are dynamic and take in dynamic lists and return dynamic
values. We may be able to cleverly assign types in the future, but the
default behaviour will always be to act on dynamic values (at least as a
fallback).

This conformance test ensures that it works in each target language, but
sadly there are currently errors in Golang. These are tracked with
#18423.

Python and Nodejs work out of the box, Go's error is not that it cant
support it but there seems to be a codegen bug