add more info when command installation fails (#18549)
When the command installation fails, we currently include the error and
stdout. `exec.ExitError` also contains the stderr of the command, which
is often more relevant when there is a failure, but hides it away unless
it's specifically used.

Add it to the error we return from `InstallPulumiCommand`. This will
hopefully help debugging CI failures we've been dealing with today, but
also more generally should help users debug why the command fails if it
does.