Set branch name for test reports on merge (#18572)
Re-opened from https://github.com/pulumi/pulumi/pull/18495

I attempted to fix this in https://github.com/pulumi/pulumi/pull/18488,
but that didn't work.

Here's a recent run for a merge to master
https://github.com/pulumi/pulumi/actions/runs/13277574078/job/37069992311#step:44:92,
and the branch is "branch":
"gh-readonly-queue/master/pr-18424-71f2757f917e853f92e09f63e21420bda653b944"

Here's a run for the daily cron job on master
https://github.com/pulumi/pulumi/actions/runs/13278773496/job/37073210482
and of course there is is `master`. This is what's been updating the
stats in codecov.