Elide SourcePosition changes (#18579)
Fixes https://github.com/pulumi/pulumi/issues/18478

This changes SourcePosition to not be considered an change that requires
an immediate snapshot write. This lets us elide the actual snapshot
write for each source position change until either a "must write" change
happens, or the end of the deployment.

Nothing in the system uses source position for correctness, so if pulumi
does crash and fails to ever write the updated source position change it
won't break anything in the system on the next run. It will just be
display metadata will be out of date for users looking in the console.
But on the next successful update they'll all get updated and attempted
to write again.

For large projects with hundreds of resources this change could save
many writes that would normally get triggered by small code changes e.g.
just shifting things by code line or folder.